+++ internal/view/singlestat_test.go
@@ -150,7 +150,7 @@ func TestSinglestatWidget(t *testing.T) {
 			test.exp(msstat)
 
 			mc := &mcontroller.Controller{}
-			mc.On("GetSingleInstantMetric", mock.Anything, test.expQuery).Return(test.controllerMetric, nil)
+			mc.On("GetSingleMetric", mock.Anything, test.expQuery, mock.Anything).Return(test.controllerMetric, nil)
 			mr := &mrender.Renderer{}
 			mr.On("LoadDashboard", mock.Anything, mock.Anything).Once().Return([]render.Widget{msstat}, nil)